+++ js/slider/slider.js
@@ -1,3 +1,5 @@
 require('./show-images.js');
 require('./next-prev.js');
+require('./more-less-pagination.js');
 require('./pagination.js');
+
